alloc: replace zk-alloc bump arena with tuned mimalloc
Remove the per-phase zk-alloc bump arena entirely and run on mimalloc, a
production allocator. The arena was fast but fragile: any allocation
outliving a phase, or a pointer retained across begin_phase's slab reset
(from a background thread, tracing, or a stray clone), silently corrupted
memory.

The arena was fast only because it reused the same pages instead of
re-faulting. We recover that with mimalloc by disabling page purging
(mi_option_purge_delay = -1, via tune_allocator()), so freed large blocks
are retained rather than returned to the OS. This matches and beats the
arena: ~11.96s vs 12.84s on `recursion --n 4` (interleaved, 3 rounds,
~7% faster than main) with none of the arena's fragility.

- delete crates/backend/zk-alloc (crate + workspace member + dep)
- src/main.rs: mimalloc global allocator (standard-alloc feature opts out)
- src/lib.rs: tune_allocator(); parallel::init() warms the pool up front
- benchmark.rs: drop begin_phase/end_phase/clone ceremony
- replace test_zk_alloc with allocator-agnostic test_aggregation
- refresh stale arena references in parallel/sumcheck doc comments
