+++ Cargo.toml
@@ -19,7 +19,6 @@ members = [
     "crates/backend/fiat-shamir",
     "crates/backend/sumcheck",
     "crates/backend/system-info",
-    "crates/backend/zk-alloc",
     "crates/backend/parallel",
 ]
 
@@ -62,7 +61,6 @@ lean_compiler = { path = "crates/lean_compiler" }
 lean_prover = { path = "crates/lean_prover" }
 rec_aggregation = { path = "crates/rec_aggregation" }
 backend = { path = "crates/backend" }
-zk-alloc = { path = "crates/backend/zk-alloc" }
 system-info = { path = "crates/backend/system-info" }
 parallel = { path = "crates/backend/parallel" }
 
@@ -84,12 +82,14 @@ include_dir = "0.7"
 
 [features]
 prox-gaps-conjecture = ["rec_aggregation/prox-gaps-conjecture"]
-standard-alloc = ["rec_aggregation/standard-alloc"]
+# Build with the plain system allocator instead of mimalloc (for comparison/debugging).
+standard-alloc = []
 
 [dependencies]
 clap.workspace = true
 rec_aggregation.workspace = true
-zk-alloc.workspace = true
+mimalloc = "0.1"
+libmimalloc-sys = { version = "0.1", features = ["extended"] }
 rand.workspace = true
 sub_protocols.workspace = true
 utils.workspace = true
