+++ crates/rec_aggregation/Cargo.toml
@@ -8,7 +8,6 @@ workspace = true
 
 [features]
 prox-gaps-conjecture = ["lean_prover/prox-gaps-conjecture"]
-standard-alloc = []
 
 [dependencies]
 utils.workspace = true
@@ -25,7 +24,6 @@ backend.workspace = true
 postcard.workspace = true
 lz4_flex.workspace = true
 serde.workspace = true
-zk-alloc.workspace = true
 
 [target.'cfg(target_os = "macos")'.dependencies]
 objc2 = { version = "0.6.4", default-features = false, features = ["std"] }
