+++ crates/rec_aggregation/src/benchmark.rs
@@ -395,9 +395,6 @@ fn build_aggregation(
     let mut last_result: Option<TypeOneMultiSignature> = None;
     let own_display_index = display_index + count_nodes(topology) - 1;
     for _ in 0..repeat {
-        #[cfg(not(feature = "standard-alloc"))]
-        zk_alloc::begin_phase();
-
         let time = Instant::now();
         let result = aggregate_type_1(
             &children,
@@ -408,13 +405,6 @@ fn build_aggregation(
         )
         .unwrap();
         let elapsed = time.elapsed();
-
-        // Clone the outputs out of the arena before the next phase resets its slabs.
-        #[cfg(not(feature = "standard-alloc"))]
-        let result = {
-            zk_alloc::end_phase();
-            result.clone()
-        };
 
         times.push(elapsed.as_secs_f64());
         last_result = Some(result);
