+++ src/lib.rs
@@ -9,8 +9,26 @@ pub use xmss::{MESSAGE_LEN_FE, XmssPublicKey, XmssSecretKey, XmssSignature, xmss
 
 pub type F = KoalaBear;
 
+/// Tune the default (mimalloc) allocator for the prover's churn of huge short-lived buffers.
+///
+/// Disables purging so freed large blocks are **retained** rather than returned to the OS
+/// and re-faulted on the next allocation. This is what made the old bump arena fast (it
+/// reused the same pages); mimalloc-with-retention matches *and beats* it here, without the
+/// arena's fragility. Idempotent; call before any heavy proving allocation.
+pub fn tune_allocator() {
+    // mimalloc v3 option index `mi_option_purge_delay` = 15; value -1 = never purge
+    // (equivalent to `MIMALLOC_PURGE_DELAY=-1`). No-op under the `standard-alloc` (plain
+    // system allocator) build.
+    #[cfg(not(feature = "standard-alloc"))]
+    unsafe {
+        libmimalloc_sys::mi_option_set(15, -1);
+    }
+}
+
 /// Call once before proving. Compiles the aggregation program and precomputes DFT twiddles.
 pub fn setup_prover() {
+    tune_allocator();
+    parallel::init(); // construct the thread pool up front (was done by `zk_alloc::begin_phase`)
     rec_aggregation::init_aggregation_bytecode();
     precompute_dft_twiddles::<F>(1 << 24);
 }
@@ -19,16 +37,3 @@ pub fn setup_prover() {
 pub fn setup_verifier() {
     rec_aggregation::init_aggregation_bytecode();
 }
-
-/// Bump-arena allocator.
-///
-/// **Optional.**
-///
-/// To enable, set it as the `#[global_allocator]` in your binary and call
-/// [`init_allocator`] once at startup. Then bracket each proving call with
-/// [`begin_phase`] / [`end_phase`] and **clone the outputs after
-/// [`end_phase`]** so the cloned copy lands in the system allocator before the
-/// next [`begin_phase`] resets the arena slabs.
-///
-/// See `tests/test_zk_alloc.rs` for a runnable end-to-end example.
-pub use zk_alloc::{ZkAllocator, begin_phase, end_phase, init as init_allocator};
